### 2023-05-06 #5449 @danielskatz
@lukeolson - thanks for your submission.  Please add the missing DOI (found above) to the bib file.  It also appears that the entry two about that one should have a DOI as well: 10.1137/18M119376

Finally, your submission message implies a previous version of this software (or a paper about it) has been published. Can you explain this a bit more, along with highlighting what is new since then, either in code, perhaps as a git diff, or just in words?
