### 2023-05-19 5ab16a5
add doi/url to bib entries
