### 2023-05-19 63a27ff
add doi/url to bib entries
