+++ docs/paper/paper.bib
@@ -208,36 +208,40 @@
 }
 
 @article{notay:2010,
-  title={An aggregation-based algebraic multigrid method},
-  author={Notay, Yvan},
-  journal={Electronic transactions on numerical analysis},
-  volume={37},
-  number={6},
-  pages={123--146},
-  year={2010}
+  title   = {An aggregation-based algebraic multigrid method},
+  author  = {Notay, Yvan},
+  journal = {Electronic transactions on numerical analysis},
+  volume  = {37},
+  number  = {6},
+  pages   = {123--146},
+  year    = {2010},
+  url     = {https://etna.math.kent.edu/vol.37.2010/pp123-146.dir/pp123-146.html},
 }
 
 @article{air1,
-  title={Nonsymmetric algebraic multigrid based on local approximate ideal restriction ($\ell$ AIR)},
-  author={Manteuffel, Thomas A and Ruge, John and Southworth, Ben S},
-  journal={SIAM Journal on Scientific Computing},
-  volume={40},
-  number={6},
-  pages={A4105--A4130},
-  year={2018},
-  publisher={SIAM}
+  title     = {Nonsymmetric algebraic multigrid based on local approximate ideal restriction ($\ell$ AIR)},
+  author    = {Manteuffel, Thomas A and Ruge, John and Southworth, Ben S},
+  journal   = {SIAM Journal on Scientific Computing},
+  volume    = {40},
+  number    = {6},
+  pages     = {A4105--A4130},
+  year      = {2018},
+  publisher = {SIAM},
+  doi       = {10.1137/17M1144350},
+  URL       = {https://doi.org/10.1137/17M1144350},
+  eprint    = {https://doi.org/10.1137/17M1144350},
 }
 
 @article{air2,
-  title={Nonsymmetric reduction-based algebraic multigrid},
-  author={Manteuffel, Thomas A and Münzenmaier, Steffen and Ruge, John and Southworth, Ben},
-  journal={SIAM Journal on Scientific Computing},
-  volume={41},
-  number={5},
-  pages={S242--S268},
-  year={2019},
-  publisher={SIAM}
+  title     = {Nonsymmetric reduction-based algebraic multigrid},
+  author    = {Manteuffel, Thomas A and Münzenmaier, Steffen and Ruge, John and Southworth, Ben},
+  journal   = {SIAM Journal on Scientific Computing},
+  volume    = {41},
+  number    = {5},
+  pages     = {S242--S268},
+  year      = {2019},
+  publisher = {SIAM},
+  doi       = {10.1137/18M1193761},
+  URL       = {https://doi.org/10.1137/18M1193761},
+  eprint    = {https://doi.org/10.1137/18M1193761},
 }
-
-
-
