### 2023-05-19 #5449 @lukeolson
> @lukeolson - thanks for your submission. Please add the missing DOI (found above) to the bib file. It also appears that the entry two about that one should have a DOI as well: 10.1137/18M119376

Fixed.  https://github.com/pyamg/pyamg/commit/5ab16a5a1fba9b9f73ea5ecf54e70b6229619ab9

> Finally, your submission message implies a previous version of this software (or a paper about it) has been published. Can you explain this a bit more, along with highlighting what is new since then, either in code, perhaps as a git diff, or just in words?

The previous version, published in JOSS, covered v4.  v5 of the code includes a series hefty PRs that were several years in development and represent major release. The new version adds these main pieces to the code.

- support for non-symmetric systems: previous solvers were limited to simply
    building a different restriction operator on A.T.  In v5, approximate ideal restriction (AIR) solvers have been added with examples.  PRs https://github.com/pyamg/pyamg/pull/367 https://github.com/pyamg/pyamg/pull/380

- complete CF AMG: v4 included only the basics of ruge-stuben (or CF) AMG.  This version adds a full suite of interpolation operators,  block support, C/F relaxation routines, etc, bringing CF-AMG up to a fully featured solver like SA-AMG in v4. PRs https://github.com/pyamg/pyamg/pull/353 https://github.com/pyamg/pyamg/pull/354 https://github.com/pyamg/pyamg/pull/244

- new pairwise aggregation: v5 adds a new pairwise solver and supporting methods; PR https://github.com/pyamg/pyamg/pull/360

- the full diff can be found here: https://github.com/pyamg/pyamg/compare/v4.2.3...v5.0.0

We feel that v5 significally expands the suite of solvers in PyAMG in comparison to v4.  This motivated the release to a new major version number and a resubmission of the article.
