### 2023-05-19 #5449 @lukeolson
> Thanks for submitting your paper to JOSS @lukeolson. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @lukeolson if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

mayrmt and mattmartineau provided reviews last time which may be efficient.  From the list, victorapm may also be familiar with solvers
