### 2023-05-24 bf56195
adding a bit more flexibility for the init sparsity pattern in AIRplus
