### 2023-05-26 a873720
Add unit tests for normal cf and fs gauss seidel
