### 2023-05-28 #5495 @editorialbot
[REVIEW]: PyAMG: Algebraic Multigrid Solvers in Python
**Submitting author:** @lukeolson (<a href="http://orcid.org/0000-0002-5283-6104">Luke Olson</a>)
**Repository:** https://github.com/pyamg/pyamg
**Branch with paper.md** (empty if default branch): joss
**Version:** v5.0.1
