### 2023-06-01 #5495 @victorapm
Hi @lukeolson, in the example section, I'm wondering if it makes more sense to show a smaller test case, maybe
```
A = pyamg.gallery.poisson((1000,1000), format='csr')
```
which would produce a test case with 1M unknowns. The reason is that's much faster to reproduce.

Along the same line, would it make sense to show at least one example featuring one of the new capabilities of pyAMG? This is not a request, I just wanted to bring up the idea...

Thank you!!
