### 2023-06-06 #5495 @lukeolson
> HI @lukeolson thanks for the replies! The `pyamg-examples` repo looks very interesting. I think it's worth and sufficient to mention it in the examples section of the paper. Thanks!

Ok, great... I updated the paper with a short note, adding that link.
