### 2023-06-06 #5495 @victorapm
Thanks, Luke! Minor detail: this needs to be updated with `1M unknowns`: https://github.com/pyamg/pyamg/blob/18d3f358dfdd406ba312a159248788ae9eafc407/docs/paper/paper.md?plain=1#L149
