### 2023-06-10 #5495 @mayrmt
@lukeolson, here are some remarks on the paper (state as of [this comment](https://github.com/openjournals/joss-reviews/issues/5495#issuecomment-1579404972):

- Line 20 seems to contain the word "among" too often: "PyAMG provides `among` the most comprehensive suite of various AMG solvers ..."
- Line 113 states, that the hierarchy in the Poisson example contains 9 levels, however the screen output in lines 100 - 112 only lists 7 levels.
