### 2023-06-10 #5495 @mayrmt
@jedbrown I worked through the manuscript as well as the software. I have raised two minor items in the manuscript text (see [my comment above](https://github.com/openjournals/joss-reviews/issues/5495#issuecomment-1585810152), which are straightforward to fix. Other than that, I don't have any further remarks. Overall, this contribution looks good and ready for publication in JOSS.
