### 2023-06-14 #5495 @lukeolson
> I second these statements. I would also add that the [CONTRIBUTING.md](https://github.com/pyamg/pyamg/blob/425f80070eefc7f4fcc6246a3b6f815f5fdbb507/CONTRIBUTING.md) page could have more detailed guidelines regarding your preferred process for issues and pull-requests for public contributions. Ex: Link to an existing issue in the PR, or if it's a new issue/feature/fix to first create an issue ticket before making a PR. Could have each issue title begin with "[Bug]", "[Fix]", "[Feature Request]", etc to help sort them quickly before your team tags them once they're posted. A quick **Guidelines** blurb with this info before the contribution assumptions would be sufficient I think.
>

Agreed, this would be helpful.  I updated the `CONTRIBUTING.md` doc with a guidelines section.  It's minimal, but good to have nonetheless.  We have tried to use consistent commit messages, but it's far from 100%.  For now I've added `should attempt to`...
