### 2023-06-14 #5495 @lukeolson
>

> * Line 20 seems to contain the word "among" too often: "PyAMG provides `among` the most comprehensive suite of various AMG solvers ..."

Agree... I simplified this sentence.

> * Line 113 states, that the hierarchy in the Poisson example contains 9 levels, however the screen output in lines 100 - 112 only lists 7 levels.

Good catch -- thanks!
