### 2023-06-21 dd9db4b
joss: minor bib fixes
