+++ docs/paper/paper.bib
@@ -76,7 +76,7 @@
 
 @book{mgtutorial,
   author    = {Briggs, William L. and Henson, Van Emden and McCormick, Steve F.},
-  title     = {A Multigrid Tutorial, Second Edition},
+  title     = {A Multigrid Tutorial},
   publisher = {Society for Industrial and Applied Mathematics},
   year      = {2000},
   doi       = {10.1137/1.9780898719505},
@@ -154,7 +154,6 @@
                   and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
   title        = {{PETS}c {W}eb page},
   url          = {https://petsc.org/},
-  howpublished = {\url{https://petsc.org/}},
   year         = {2021},
 }
 
