+++ README.md
@@ -50,11 +50,11 @@ Portions of the project were partially supported by the NSF under award DMS-0612
 If you use PyAMG in your work, please consider using the following citation:
 
 <pre>
-@article{BeOlSc2023,
+@article{BeOlScSo2023,
   author    = {Nathan Bell and Luke N. Olson and Jacob Schroder and Ben S. Southworth},
   title     = {{PyAMG}: Algebraic Multigrid Solvers in Python},
   journal   = {Journal of Open Source Software},
-  year      = {2022},
+  year      = {2023},
   publisher = {The Open Journal},
   note      = {submitted},
 }
