+++ CITATION.bib
@@ -1,8 +1,8 @@
-@article{BeOlSc2023,
+@article{BeOlScSo2023,
   author    = {Nathan Bell and Luke N. Olson and Jacob Schroder and Ben S. Southworth},
   title     = {{PyAMG}: Algebraic Multigrid Solvers in Python},
   journal   = {Journal of Open Source Software},
-  year      = {2022},
+  year      = {2023},
   publisher = {The Open Journal},
   note      = {submitted},
 }
