### 2023-06-28 a56540b
converting to CSR so filter_operator_rows works recursively
