### 2023-06-28 ed93ef9
going back on CSR matrix here...not sure how, but this affects other routines in addition ot filter_matrix_rows
