### 2023-07-03 #5495 @lukeolson
>

> Thanks for the excellent timely reviews! I submitted a small PR on the paper. Assuming no issues merging that, @lukeolson, please proceed with these final steps:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.

Ok, here we go:
- v5.0.1 (https://github.com/pyamg/pyamg/releases/tag/v5.0.1)
- Zenodo https://zenodo.org/record/8109414
- The metadata looks correct.
- Zenodo doi: 10.5281/zenodo.8109414

Let me know if anything else is needed.  Thanks @jedbrown for overseeing this -- and thanks again to the referees for their time.
