### 2023-07-03 f247844
Slight update to the CLAIR solver since paper submission
