### 2023-07-03 12e8e31
removing the hacked version of fcf that was hacked to mimic l1 FF jacobi -- fcf should now be fcf jacobi again
