### 2023-07-04 #5495 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
