+++ docs/paper/paper.bib
@@ -77,7 +77,7 @@
 @book{mgtutorial,
   author    = {Briggs, William L. and Henson, Van Emden and McCormick, Steve F.},
   title     = {A Multigrid Tutorial},
-  publisher = {Society for Industrial and Applied Mathematics},
+  publisher = {{Society for Industrial and Applied Mathematics}},
   year      = {2000},
   doi       = {10.1137/1.9780898719505},
   address   = {},
@@ -131,15 +131,15 @@
 }
 
 @manual{trilinos-website,
-  title  = {The {T}rilinos {P}roject {W}ebsite},
-  author = {The {T}rilinos {P}roject {T}eam},
+  title  = {{T}rilinos {P}roject Website},
+  author = {{Trilinos Project Team}},
   year   = {2020},
   url    = {https://trilinos.github.io}
 }
 
 @manual{muelu-website,
-  title  = {The {M}ue{L}u {P}roject {W}ebsite},
-  author = {The {M}ue{L}u {P}roject {T}eam},
+  title  = {{M}ue{L}u {P}roject Website},
+  author = {{MueLu Project Team}},
   year   = {2020},
   url    = {https://trilinos.github.io/muelu.html}
 }
