### 2023-07-04 #5495 @danielskatz
@lukeolson - I've proofread the paper, and have a few small suggestions for the bib, in https://github.com/pyamg/pyamg/pull/391 - Please merge this, or let me know what you disagree with, then we can proceed to accept and publish this work
